them bang product , con phai doi du lieu de render ra
