+++ src/main/java/com/nhom36/milkPowder/controllers/adminController/ProductsController.java
@@ -1,0 +1,19 @@
+package com.nhom36.milkPowder.controllers.adminController;
+
+import javax.servlet.*;
+import javax.servlet.http.*;
+import javax.servlet.annotation.*;
+import java.io.IOException;
+
+@WebServlet(name = "ProductsController", value = "/ProductsController")
+public class ProductsController extends HttpServlet {
+    @Override
+    protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
+    doPost(request,response);
+    }
+
+    @Override
+    protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
+        request.getRequestDispatcher("admin/products.jsp").forward(request,response);
+    }
+}
