--- src/main/java/com/nhom36/milkPowder/controllers/adminController/TableController.java
@@ -1,19 +1,0 @@
-package com.nhom36.milkPowder.controllers.adminController;
-
-import javax.servlet.*;
-import javax.servlet.http.*;
-import javax.servlet.annotation.*;
-import java.io.IOException;
-
-@WebServlet(name = "TableController", value = "/TableController")
-public class TableController extends HttpServlet {
-    @Override
-    protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-    doPost(request,response);
-    }
-
-    @Override
-    protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-        request.getRequestDispatcher("admin/tables.jsp").forward(request,response);
-    }
-}
