+++ src/main/webapp/admin/setup/head.jsp
@@ -1,0 +1,24 @@
+<%--
+  Created by IntelliJ IDEA.
+  User: hoanghai
+  Date: 2023-01-05
+  Time: 12:37 AM
+  To change this template use File | Settings | File Templates.
+--%>
+<%@page contentType="text/html; charset=UTF-8" language="java" pageEncoding="UTF-8" %>
+<%@ taglib prefix="c" uri="http://java.sun.com/jsp/jstl/core" %>
+
+  <meta charset="utf-8">
+  <meta http-equiv="X-UA-Compatible" content="IE=edge">
+  <meta name="viewport" content="width=device-width, initial-scale=1">
+  <title>HachaMilk-Admin</title>
+
+  <!-- Tailwind is included -->
+
+  <link rel="apple-touch-icon" sizes="180x180" href="admin/img/logo.png"/>
+  <link rel="icon" type="image/png" sizes="32x32" href="admin/img/logo.png"/>
+  <link rel="icon" type="image/png" sizes="16x16" href="admin/img/favicon-16x16.png"/>
+  <link rel="stylesheet" href="admin/css/main.css">
+  <link rel="stylesheet" href="https://cdn.datatables.net/1.11.3/css/dataTables.bootstrap5.min.css">
+
+
