--- src/main/webapp/admin/view/footer.jsp
@@ -1,16 +1,0 @@
-<%--
-  Created by IntelliJ IDEA.
-  User: hoanghai
-  Date: 2023-01-05
-  Time: 12:02 AM
-  To change this template use File | Settings | File Templates.
---%>
-<%@ page contentType="text/html;charset=UTF-8" language="java" %>
-<html>
-<head>
-    <title>Title</title>
-</head>
-<body>
-
-</body>
-</html>
