+++ src/main/webapp/admin/view/footer_admin.jsp
@@ -1,0 +1,27 @@
+<%--
+  Created by IntelliJ IDEA.
+  User: hoanghai
+  Date: 2023-01-05
+  Time: 12:02 AM
+  To change this template use File | Settings | File Templates.
+--%>
+<%@page contentType="text/html; charset=UTF-8" language="java" pageEncoding="UTF-8" %>
+<%@ taglib prefix="c" uri="http://java.sun.com/jsp/jstl/core" %>
+<footer class="footer">
+    <div class="flex flex-col md:flex-row items-center justify-between space-y-3 md:space-y-0">
+        <div class="flex items-center justify-start space-x-3">
+            <div>
+                © 2021, nhom36
+            </div>
+
+            <div>
+                <p>Distributed By: <a href="https://github.com/NgHgHai/Nhom36_powderMilk_Ecommerce" target="_blank">Nhóm
+                    36</a></p>
+            </div>
+            <a href="https://github.com/justboil/admin-one-tailwind" style="height: 20px">
+                <img src="https://img.shields.io/github/v/release/justboil/admin-one-tailwind?color=%23999">
+            </a>
+        </div>
+
+    </div>
+</footer>
